FIX bug

+++ challenge-01/calc_area_circle.py
@@ -1,6 +1,6 @@
 # Calculate area of a circle
 PI = 3.14
 radius = input("รัศมีวงกลมเท่าไหร่? (cm) ")
-radius = int(radius)
+radius = float(radius)
 area_of_circle = PI * radius ** 2
 print("พื้นที่วงกลมคือ " + str(area_of_circle) + " ตารางเซนติเมตร")
