+++ challenge-01/calc_area_triangle.py
@@ -1,7 +1,7 @@
 # Caluclate area of a triangle
 base = input("ความยาวฐานของสามเหลี่ยมเท่าไหร่? (cm) ")
 height = input("ความสูงของสามเหลี่ยมเท่าไหร่? (cm) ")
-base = int(base)
-height = int(height)
+base = float(base)
+height = float(height)
 area_of_triangle = 0.5 * base * height
 print("พื้นที่ของสามเหลี่ยมคือ " + str(area_of_triangle) + " ตารางเซนติเมตร")
